Allow variable type to be overridden. #29.
MDES 3.3 has a table with several PKs according to the VDR schema
(link_child_partic_org). Upon inspection, it is plain that the extras
are supposed to be FKs.

This feature allows overriding the named type provided for a variable
in the schema to allow switching these to be FKs.